separated graph module from app for coretk
